% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaness_classifier}
\alias{brcaness_classifier}
\title{Random Forest Classifier for BRCAness}
\format{
A list containing a trained random forest classifier object.
}
\usage{
data(brcaness_classifier)
}
\description{
This data object contains a trained random forest classifier that utilizes a
24-gene expression signature to classify patients into two categories: those
with BRCAness and those with BRCA status. The classifier was trained using
multiomics data, including RNA-Seq data, from the TCGA-OV dataset.
}
\details{
A trained random forest classifier for predicting BRCAness status using a
24-gene expression signature. This classifier was trained with multiomics
data from the TCGA-OV dataset and utilizes RNA sequencing (RNA-Seq) data to
classify patients as either having BRCAness or BRCA status.


The random forest classifier in this data object was trained with feature
selection to optimize its ability to predict BRCAness status based on RNA-Seq
data. It is a result of a multiomics approach aimed at accurately classifying
patients.
}
\examples{
# Load the trained BRCAness classifier
data(brcaness_classifier)

# Predict BRCAness status for a new patient using the classifier
new_patient_data <- ...  # Replace with new patient's RNA-Seq data
prediction <- predict(brcaness_classifier, new_patient_data)

}
