% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brcaness.R
\name{BRCAness_immunotype}
\alias{BRCAness_immunotype}
\title{Stratify HGSOC Patients Based on BRCAness and Immune Subtypes}
\usage{
BRCAness_immunotype(se)
}
\arguments{
\item{se}{A `SummarizedExperiment` object containing patient data, including BRCAness status, tumor immunephenotype, and molecular subtype.}
}
\value{
The modified `SummarizedExperiment` object with an added `immunotype` column in `colData`, classifying each patient as 'BRIT', 'noBRIT', or 'other'.
}
\description{
This function stratifies patients in a `SummarizedExperiment` object into categories based on BRCAness status, tumor immunephenotype, and molecular subtype. It identifies patients with a BRCAness immunetype (BRIT), which combines the BRCAness phenotype with an infiltrated tumor immune phenotype and an immune-reactive molecular subtype (IMR).
}
\details{
The function uses a list of 157 genes and random forest analysis to classify the tumor immunephenotype of each patient as infiltrated, excluded, or desert. The BRIT group consists of patients with BRCAness, an infiltrated tumor immune phenotype, and an immune-reactive molecular subtype (IMR). This group is potentially more responsive to PARP inhibitor and immune checkpoint inhibitor therapies. The function also accounts for the presence of suppressive immune cells and does not show a significant difference in overall survival between BRIT and noBRIT groups. This stratification provides insights into the complex interplay between genetic predisposition and immune response in HGSOC, aiding in personalized treatment planning.
}
\examples{
# se is a pre-loaded SummarizedExperiment object
se <- BRCAness_immunotype(se)

}
