% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunophenoscore.R
\name{calculateIPS}
\alias{calculateIPS}
\title{Calculate Immunophenoscore (IPS) from a SummarizedExperiment}
\usage{
calculateIPS(se)
}
\arguments{
\item{se}{A `SummarizedExperiment` object containing normalized gene expression data.}
}
\value{
A data frame with samples as rows and calculated scores (WG, MHC, CP, EC, SC, MDSC, TREG, AZ, IPS) as columns.
}
\description{
This function calculates the Immunophenoscore (IPS) and its components scores
from gene expression data encapsulated in a `SummarizedExperiment` object.
It requires a specific set of genes and corresponding weights provided in a separate file.
}
\references{
Immunophenogram:
This R-script can be used to calculate Immunophenoscore (IPS) and generate Immunophenogram
from "EXPR.txt" and "IPS_genes.txt". The script and associated documentation can be found at
the following URL: https://github.com/icbi-lab/Immunophenogram
(C) ICBI, Medical University of Innsbruck, Biocenter, Division of Bioinformatics
Version 1.0 08.07.2016
}
