% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{classifier_infiltration_status}
\alias{classifier_infiltration_status}
\title{Tumor Infiltration Status Classifier}
\format{
An object of class `randomForest` (inherits from `list`), representing
  a fitted Random Forest model.
}
\usage{
data(classifier_infiltration_status)
}
\description{
A Random Forest classifier trained on gene expression data from the ICON7 trial.
This classifier is designed to classify the tumor immune infiltration status
into categories such as 'infiltrated', 'excluded', or 'desert' based on a
gene signature. The model is trained using `randomForest` package and can be
used to predict the infiltration status of ovarian cancer samples.
}
\details{
The classifier was trained using a Random Forest algorithm with 300 trees
on the ICON7 dataset, which contains gene expression data for ovarian cancer
samples. The gene signature used for the training consists of genes identified
as common between the SummarizedExperiment datasets and the tumor immune phenotype
signature.

The `classifier_infiltration_status` is saved as an R object of class `randomForest`,
which contains the entire fitted model object. This model can be used to predict
infiltration status in other datasets by supplying the appropriate gene expression
data for the genes included in the signature.

The classifier should be applied only to data that has been preprocessed in the same
manner as the ICON7 trial data to ensure the validity of the predictions.
}
\examples{
data(classifier_infiltration_status)
# Suppose `new_data` is a matrix of gene expression values with rows as genes and columns as samples:
predicted_status <- predict(classifier_infiltration_status, new_data)
}
\references{
Desbois M, Udyavar AR, Ryner L, Kozlowski C, Guan Y, Dürrbaum M, et al. Integrated digital
pathology and transcriptome analysis identifies molecular mediators of T-cell exclusion in
ovarian cancer. Nat Commun. 2020;11:5583.
}
