% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{classify_brcaness}
\alias{classify_brcaness}
\title{Classify Samples Using BRCAness Classifier}
\usage{
classify_brcaness(se, brcaness_classifier = NA, brcaness_signature = NA)
}
\arguments{
\item{se}{A SummarizedExperiment object with count data.}

\item{brcaness_classifier}{A random forest classifier object trained on
BRCAness samples.}

\item{brcaness_signature}{A vector of gene symbols representing the BRCAness
gene signature.}
}
\value{
A SummarizedExperiment with BRCAness predictions, in ColData.
}
\description{
This function uses the BRCAness classifier to predict the BRCAness status of
the samples in a SummarizedExperiment object. It first checks that all the
genes in the BRCAness signature are present in the count data of the
SummarizedExperiment object before making predictions.
}
