% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration_status.R
\name{classify_infiltration_status}
\alias{classify_infiltration_status}
\title{Classify Tumor Immune Infiltration Status}
\usage{
classify_infiltration_status(se, classifier, gene_signature)
}
\arguments{
\item{se}{A SummarizedExperiment object containing gene expression data where samples are
columns and genes are rows.}

\item{classifier}{A trained classifier object capable of making predictions based on the
gene expression data provided in `se`.}

\item{gene_signature}{A character vector containing gene symbols that make up the gene
signature used by the classifier. The function will subset the expression data in `se`
based on these genes to make predictions.}
}
\value{
The SummarizedExperiment object `se` with an additional column in `colData` named
  `InfiltrationStatus`, which contains the classification results for each sample.
}
\description{
This function applies a trained classifier to a SummarizedExperiment object to classify
samples based on their tumor immune infiltration status using a specified gene signature.
}
\details{
The function first validates that the provided `se` object is indeed a SummarizedExperiment.
It then extracts the expression data for the genes in the `gene_signature` from the `se` object.
This gene expression matrix is transposed (samples as rows, genes as columns) and used as input
for the classifier to predict the infiltration status. The predictions are then added to the
`colData` of the `se` object, allowing for easy retrieval and further analysis.
}
\examples{
# Assuming `se` is a SummarizedExperiment object with gene expression data,
# `rf_classifier` is a trained random forest classifier, and
# `immune_genes` is a vector of gene symbols making up the immune signature:
se <- classify_infiltration_status(se, rf_classifier, immune_genes)

}
