% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{compute_angiogenesis_score}
\alias{compute_angiogenesis_score}
\title{Compute Angiogenesis Score for SummarizedExperiment Data}
\usage{
compute_angiogenesis_score(se)
}
\arguments{
\item{se}{`SummarizedExperiment` object containing the necessary gene expression data.}

\item{sample_id}{The identifier for the sample to compute the score.}
}
\value{
`SummarizedExperiment` object with the new `angiogenesis_score` column added to its `colData`.
}
\description{
Calculates an angiogenesis score based on expression levels of seven angiogenic genes. This score ranges from 0 to 3 and is added to the `colData` of the `SummarizedExperiment` object.
}
\details{
The function evaluates the expression of VEGFA, VEGFR2, PDGFA, PDGFB, PDGFRA, PDGFRB, and KIT.
It dichotomizes these based on a predefined threshold and calculates a score:
- 0 for 0 genes with high expression
- 1 for 1-3 genes with high expression
- 2 for 4-5 genes with high expression
- 3 for 6-7 genes with high expression
This score is indicative of the angiogenic potential of the tumor and has been linked to patient prognosis and response to angiogenesis inhibitors like bevacizumab.
}
\examples{
# se is a pre-loaded SummarizedExperiment object
sample_id <- "sample123"
se <- compute_angiogenesis_score(se, sample_id)

}
\references{
Wieser V, Tsibulak I, Reimer DU, Zeimet AG, Fiegl H, Hackl H, Marth C. An angiogenic tumor phenotype
predicts poor prognosis in ovarian cancer. Gynecol Oncol. 2023 Mar;170:290-299.
doi: 10.1016/j.ygyno.2023.01.034. Epub 2023 Feb 7. PMID: 36758419.
}
