% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{computeCytC1qaRatio}
\alias{computeCytC1qaRatio}
\title{Compute CYT to C1QA Ratio (C2C)}
\usage{
computeCytC1qaRatio(se)
}
\arguments{
\item{se}{A SummarizedExperiment object containing gene expression data.
The object must contain expression data for GZMB, PRF1, and C1QA genes.}
}
\value{
Returns the SummarizedExperiment object with an additional column
        in `colData` named `ratio_CYT_C1QA`, representing the computed C2C ratio.
}
\description{
This function calculates the cytolytic activity (CYT) to C1QA ratio (C2C)
for a given SummarizedExperiment object. The ratio is calculated using the
expression values of GZMB, PRF1, and C1QA. The C2C ratio is then added
to the `colData` of the provided SummarizedExperiment object.
}
\details{
The function computes the C2C ratio using the formula:
         C2C = 0.5 x (log2(TPM + 1 of GZMB) + log2(TPM + 1 of PRF1)) / log2(TPM + 1 of C1QA).
         It assumes that the expression data in the `SummarizedExperiment` object
         are either TPM (Transcripts Per Million) or intensity values suitable for
         the log transformation.
}
\examples{
# se is a SummarizedExperiment object with expression data for GZMB, PRF1, and C1QA
updatedSe <- computeCytC1qaRatio(se)

}
