% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{computeVulnerabilityScore}
\alias{computeVulnerabilityScore}
\title{Compute Vulnerability Score for SummarizedExperiment Data}
\usage{
computeVulnerabilityScore(se)
}
\arguments{
\item{se}{A `SummarizedExperiment` object that contains or can compute the BRCAness probability and the C2C ratio.}
}
\value{
The input `SummarizedExperiment` object with added columns for the mapped C2C ratio and the computed vulnerability score.
}
\description{
This function calculates the vulnerability score for each sample in a `SummarizedExperiment` object. It requires the BRCAness probability and the ratio of cytolytic activity to C1QA (C2C ratio) as part of the dataset. If these are not present, the function computes them. The vulnerability score is computed using specified coefficients and a transformation function.
}
\details{
The function first checks if the `SummarizedExperiment` object contains `BRCAness_Prob` and `ratio_CYT_C1QA`. If not, it computes these using `classify_brcaness` and `computeCytC1qaRatio` functions, respectively. The vulnerability score is calculated using a logistic transformation of the C2C ratio and a linear combination of this transformed value with the BRCAness probability, using predefined coefficients. This score aims to reflect the patient's vulnerability based on their molecular profile.
}
\examples{
# se is a pre-loaded SummarizedExperiment object
se <- computeVulnerabilityScore(se)

}
