% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_subtype.R
\name{get_consensus_ov_subtypes}
\alias{get_consensus_ov_subtypes}
\title{Obtain Consensus Subtypes of High-Grade Serous Ovarian Cancer}
\usage{
get_consensus_ov_subtypes(
  se,
  ids_type = "symbol",
  concordant.tumors.only = TRUE,
  remove.using.cutoff = FALSE,
  percentage.dataset.removed = 0.75
)
}
\arguments{
\item{se}{A Summarized Experiments of gene expression values, with gene symbol or entrez id as identifiers.}

\item{ids_type}{A character string indicating the type of IDs used in the 'data' row names. Can be either "symbol" or "entrez".}

\item{concordant.tumors.only}{Logical. If TRUE, only tumors that are concordantly classified across all datasets are included in the final classification. Defaults to TRUE.}

\item{remove.using.cutoff}{Logical. If TRUE, tumors with poor classification confidence are removed using the optimal cutoff. Defaults to FALSE.}

\item{percentage.dataset.removed}{Numeric value indicating the percentage of the dataset to be removed based on classification confidence. Defaults to 0.75.}
}
\value{
Summarized Experiments with Tumor_Molecular_Subtypes of consensus molecular subtypes for the samples, in colData.
}
\description{
This function gets the consensus molecular subtypes of high-grade serous (HGS) ovarian cancer using the `consensusOV` package.
The function implements a consensus classifier which consolidates and improves on the robustness of proposed subtype classifiers.
}
\examples{
load_TCGA_OV
TCGA_OV <- get_consensus_ov_subtypes(TCGA_OV, ids_type = "symbol")
}
\references{
Chen G, Kannan L, Geistlinger L, Kofia V, Safikhani Z, Gendoo D, Parmigiani G, Birrer M, Haibe-Kains B, Waldron L (2018).
"Consensus on molecular subtypes of high-grade serous ovarian carcinoma."
Clinical Cancer Research, 24, 4990. doi:10.1158/1078-0432.CCR-18-0784.
}
