% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{getEnsemblIds}
\alias{getEnsemblIds}
\title{Get Ensembl IDs from Gene Symbols}
\usage{
getEnsemblIds(gene_symbols, org = "hsapiens_gene_ensembl")
}
\arguments{
\item{gene_symbols}{A character vector of gene symbols to look up}

\item{org}{A string specifying the organism (default: "hsapiens_gene_ensembl")}
}
\value{
A data frame with two columns: "ensembl_gene_id" and "external_gene_name"
}
\description{
Takes a vector of gene symbols and returns a data frame of Ensembl gene IDs
and gene symbols using the biomaRt package.
}
\examples{
getEnsemblIds(c("BRCA1", "TP53"))
}
