% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{immune_signature_score}
\alias{immune_signature_score}
\title{Calculate enrichment scores for immune signatures}
\usage{
immune_signature_score(se, method = "ssgsea", genelist)
}
\arguments{
\item{se}{A SummarizedExperiment object containing RNA sequencing data}

\item{method}{A character string indicating the method to use for calculating enrichment scores. Default is "ssgsea".}

\item{genelist}{A character vector with the gene list for the immune signature.}
}
\value{
A SummarizedExperiment object with the enrichment scores added to the colData.
}
\description{
This function calculates enrichment scores for a list of immune signatures
using the GSVA method. The user can provide their own gene list or use one
of the pre-defined ones. The function returns an updated SummarizedExperiment
object with the enrichment scores added to the colData.
}
