% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune_signatures}
\alias{immune_signatures}
\title{Immune Signatures}
\format{
A list of lists, where each sublist contains:
  - `Name`: The name of the immune signature.
  - `Genes`: A character vector of gene symbols representing the signature genes.
}
\usage{
data(immune_signatures)
}
\description{
The `immune_signatures` object contains well-defined immune-related gene
signatures in GMT format. These signatures are valuable for characterizing
immune processes and activities in biological data, including gene expression
profiles.
}
\details{
A GMT (Gene Matrix Transposed) file containing well-defined immune-related
gene signatures. Each signature consists of a list with the name of the
signature and a vector of gene symbols representing the genes that constitute
the signature. These signatures are used to characterize immune-related
processes, including T cell inflammation, IFN gamma signature, cytolytic
activity, and cytotoxic T lymphocyte function.
}
\examples{
# Load the immune signatures
data(immune_signatures)

# Access the genes in the "T cell inflammation" signature
t_cell_inflammation_genes <- immune_signatures$T_cell_inflammation$Genes

}
\seealso{
For more information on GMT files and gene set enrichment analysis (GSEA),
refer to the relevant literature and software documentation.
}
