% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_signature.R
\name{load_brcaness_signature}
\alias{load_brcaness_signature}
\title{Load BRCAness Gene Signature}
\usage{
load_brcaness_signature()
}
\value{
A vector with gene symbol for the BRCAness gene signature.
}
\description{
Loads the BRCAness gene signature from the "brcaness_signature.rda" data file
and returns it as a vector with gene symbol for the for 24 genes.
}
\examples{
# Load the BRCAness gene signature
brcaness_signature <- load_brcaness_signature()

# Access gene expression values for the first gene
brcaness_signature$Gene1

}
