% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration_status.R
\name{load_classifier_infiltration_status}
\alias{load_classifier_infiltration_status}
\title{Load Infiltration Status Classifier}
\usage{
classifier_infiltration_status <- load_classifier_infiltration_status()
}
\value{
A pre-trained classifier object. Typically, this is a machine learning model object
that can be used to predict infiltration status based on gene expression profiles.
}
\description{
Loads the pre-trained classifier for tumor immune infiltration status from the
"classifier_infiltration_status.rda" data file. This classifier is used to determine
the infiltration status (e.g., infiltrated, excluded, desert) within the tumor microenvironment.
}
\examples{
# Load the Infiltration Status classifier
classifier_infiltration_status <- load_classifier_infiltration_status()

# Now you can use `classifier_infiltration_status` with gene expression data to
# predict infiltration statuses.

}
