% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_signature.R
\name{load_immune_signatures}
\alias{load_immune_signatures}
\title{Load Immune Signatures}
\usage{
load_immune_signatures()
}
\value{
A list of lists, where each sublist contains:
  - `Name`: The name of the immune signature.
  - `Genes`: A character vector of gene symbols representing the signature genes.
}
\description{
Loads the immune signatures from the "immune_signatures.rda" data file and
returns them as a list of lists, where each sublist contains the name of
an immune signature and a vector of gene symbols representing the genes that
constitute the signature.
}
\examples{
# Load the immune signatures
immune_signatures <- load_immune_signatures()

# Access the genes in the "T cell inflammation" signature
t_cell_inflammation_genes <- immune_signatures$T_cell_inflammation$Genes

}
