% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_signature.R
\name{load_small_immune_signatures}
\alias{load_small_immune_signatures}
\title{Load Small Immune Signatures}
\usage{
small_immune_signatures <- load_small_immune_signatures()
}
\value{
A list of lists, where each sublist contains:
  - `Name`: The name of the small immune signature.
  - `Genes`: A character vector of gene symbols representing the genes in the signature.
}
\description{
Loads the small immune signatures from the "small_immune_signatures.rda" data file and
returns them as a list of lists. Each sublist corresponds to an immune signature that
contains fewer than 10 genes, making it suitable for analysis methods like z-score computations
that may be more appropriate for smaller gene sets than single-sample gene set variance analysis.
}
\examples{
# Load the small immune signatures
small_immune_signatures <- load_small_immune_signatures()

# Access the genes in the "T cell inflammation" signature
ifng_ayers <- small_immune_signatures$'IFNG Ayers'

}
