% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{load_TCGA_OV}
\alias{load_TCGA_OV}
\title{Load TCGA-OV Dataset}
\usage{
load_TCGA_OV()
}
\value{
A SummarizedExperiment object containing raw RNA-Seq counts from the TCGA-OV dataset
with associated metadata columns.
}
\description{
Loads the TCGA-OV dataset from the "TCGA_OV.rda" file and returns it as a
SummarizedExperiment object.
}
\examples{
# Load the TCGA-OV dataset
my_dataset <- load_TCGA_OV()

# Access metadata columns
metadata(my_dataset)$AGE
metadata(my_dataset)$TUMOR_GRADE

}
