% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{load_tcgaStats}
\alias{load_tcgaStats}
\title{Load TCGA Statistics Data}
\usage{
load_tcgaStats()
}
\value{
A data frame containing TCGA statistics.
}
\description{
This function loads the TCGA statistics data from the 'OvRSeq' package. The data is typically used for comparative analysis in the context of high-grade serous ovarian cancer (HGSOC) research.
}
\details{
The function retrieves the 'tcgaStats' dataset, which includes various statistical measures derived from The Cancer Genome Atlas (TCGA) data. This dataset is crucial for analyzing and comparing patient-specific data against a broader cancer database, facilitating insights into molecular patterns, biomarker distributions, and other critical aspects in HGSOC studies.
}
\examples{
tcga_stats <- load_tcgaStats()

}
