% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration_status.R
\name{load_tumor_immune_phenotype_signature}
\alias{load_tumor_immune_phenotype_signature}
\title{Load Tumor Immune Phenotype Signature}
\usage{
tumor_immune_phenotype_signature <- load_tumor_immune_phenotype_signature()
}
\value{
A character vector of gene symbols representing the tumor immune phenotype signature.
}
\description{
Loads the tumor immune phenotype signature from the
"tumor_immune_phenotype_signature.rda" data file. This signature contains
a list of genes associated with the immune phenotype of tumors.
}
\examples{
# Load the tumor immune phenotype signature
tumor_immune_phenotype_signature <- load_tumor_immune_phenotype_signature()

}
