% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{log2Norm}
\alias{log2Norm}
\title{Log2 normalization of count data}
\usage{
log2Norm(count_data)
}
\arguments{
\item{count_data}{A matrix of count data with genes in rows and samples in columns..}
}
\value{
A matrix of count data with genes in rows and samples in columns, containing the log2 normalized count data.
}
\description{
This function performs log2 normalization of count data. It adds 1 to each count value and takes the logarithm base 2 of the resulting value.
}
\examples{
data("example_counts")
data("example_gene_length")
se <- SummarizedExperiment::SummarizedExperiment(assays = list(counts = example_counts))
log2_norm_data <- log2Norm(se)

}
