% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OvaRSeqDataSet.R
\docType{class}
\name{OvaRSeqDataSet-class}
\alias{OvaRSeqDataSet-class}
\title{OvaRSeqDataSet class}
\description{
\code{OvaRSeqDataSet} is a class inherited from
\code{\link{SummarizedExperiment}}.
It is used to store the count matrix, colData, and design formula
in differential expression analysis.
}
\references{
Martin Morgan, Valerie Obenchain, Jim Hester and
Hervé Pagès (2018). SummarizedExperiment: SummarizedExperiment container.
R package version 1.12.0.
}
