% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OvRSeq.R
\name{OvRSeq}
\alias{OvRSeq}
\title{Comprehensive RNA Sequencing-based Characterization of HGSOC Samples}
\usage{
OvRSeq(se, normalize = FALSE)
}
\arguments{
\item{se}{A SummarizedExperiment object containing RNA-seq data for HGSOC samples.}

\item{normalize}{A logical parameter indicating whether the assay data should be normalized.
If FALSE, the function will expect raw count data in integer form.}
}
\value{
Returns a SummarizedExperiment object enriched with BRCAness classification,
        molecular subtype classification, immune phenotyping, IPS scoring, immune signatures,
        and immune cell deconvolution estimates.
}
\description{
This function provides an all-encompassing analysis of high-grade serous ovarian cancer (HGSOC)
RNA-seq data. It integrates various predictive biomarkers and performs a multi-faceted immune
profiling, leveraging a 24-gene expression signature to classify BRCAness, subtype classification,
immune environment profiling, and immune cell deconvolution.
}
\examples{
# Load data
se <- load_TCGA_OV()
# Run the OvRSeq analysis pipeline
se <- OvRSeq(se)

}
