% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ovrseq_results.R
\name{plot_deconvolution_data}
\alias{plot_deconvolution_data}
\title{Plot Deconvolution Data for a Given Sample}
\usage{
plot_deconvolution_data(
  se,
  sample_id,
  deconvolution_methods = c("quantiseq", "timer", "mcp_counter")
)
}
\arguments{
\item{se}{A `SummarizedExperiment` object containing the deconvolution data in its `colData`.}

\item{sample_id}{The identifier of the sample for which the plot will be generated.}

\item{deconvolution_methods}{A vector of strings representing the deconvolution methods to check
in the `colData` of the `SummarizedExperiment` object. The function will use the first method that
has corresponding data available.}
}
\value{
A ggplot object representing the estimated immune cell infiltrate fractions.
}
\description{
Creates a bar plot representing the estimated immune cell infiltrate fractions for a specified sample
using the first available deconvolution method from a provided list.
}
\details{
The function iterates through the provided list of deconvolution methods and checks if there are
any columns in the `colData` of the `SummarizedExperiment` object corresponding to these methods.
It then uses the data from the first available method to create a bar plot. If no data is found
for any of the provided methods, the function stops with an error.
}
\examples{
# Assuming se is a pre-loaded SummarizedExperiment object and sample_id is known
immunedeconv_methods <- c("quantiseq", "timer", "mcp_counter")
sample_id <- "sample123"  # Replace with an actual sample ID
p <- plot_deconvolution_data(se, sample_id, immunedeconv_methods)
print(p)

}
