% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ovrseq_results.R
\name{plot_ggmarginal}
\alias{plot_ggmarginal}
\title{Plot Marginal Distributions for OvRSeq Results}
\usage{
plot_ggmarginal(se, x_var, y_var, color_var)
}
\arguments{
\item{se}{A SummarizedExperiment object containing results from OvRSeq.}

\item{x_var}{The name of the variable in colData(se) to use for the x-axis.}

\item{y_var}{The name of the variable in colData(se) to use for the y-axis.}

\item{color_var}{The name of the variable in colData(se) to use for coloring the points.}
}
\value{
A ggplot object with added marginal histograms.
}
\description{
This function takes the results from OvRSeq, extracts specified variables from
the column data, and creates a ggplot with marginal histograms to show the
distribution of x and y variables separately.
}
\examples{
# Assuming `se` is a SummarizedExperiment object with relevant data
# plot_ggmarginal(se, "variable1", "variable2", "groupingVar")

}
