% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ovrseq_results.R
\name{plot_ggmarginal_sample}
\alias{plot_ggmarginal_sample}
\title{Plot one sample for report}
\usage{
plot_ggmarginal_sample(
  se,
  x_var = "C1QA",
  y_var = "CD8A",
  color_var = "BRCAness"
)
}
\arguments{
\item{se}{A `SummarizedExperiment` object.}

\item{x_var}{A string representing the variable from `se` to be plotted on the x-axis (default is "C1QA").}

\item{y_var}{A string representing the variable from `se` to be plotted on the y-axis (default is "CD8A").}

\item{color_var}{A string representing the variable from `se` to be used for coloring points (default is "BRCAness_Prob").}
}
\value{
A `ggplot` object representing the scatter plot with marginal histograms.
}
\description{
This function creates a scatter plot with marginal histograms for specified variables from a `SummarizedExperiment` object. It combines the provided `se` object with the TCGA_OV dataset, plots specified variables, and highlights a specific sample from `se` with a distinct style.
}
\details{
The function first combines the provided `se` object with the TCGA_OV dataset. It then creates a scatter plot for the specified x and y variables, with points colored based on the specified `color_var`. A specific sample from `se` is highlighted in the plot with a black border and white fill, distinct from other data points.

The function assumes that the TCGA_OV dataset and `se` object have compatible structures and the required columns. The user needs to ensure that the TCGA_OV dataset is loaded and accessible.
}
\examples{
# se is a pre-loaded SummarizedExperiment object
plot_ggmarginal(se, "C1QA", "CD8A", "BRCAness_Prob")

}
