% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ovrseq_results.R
\name{plot_immune_signature_one_sample}
\alias{plot_immune_signature_one_sample}
\title{Plot Immune Signature Scores for a Single Sample}
\usage{
plot_immune_signature_one_sample(se, sample_id)
}
\arguments{
\item{se}{A `SummarizedExperiment` object containing immune-related metrics in its `colData`.}

\item{sample_id}{A character string specifying the ID of the sample to be plotted.}
}
\value{
A `ggplot` object representing the bar plot of immune signature scores for the specified sample.
}
\description{
This function generates a bar plot of selected immune signature scores for a specified sample from a `SummarizedExperiment` object. It filters the data to include specific immune-related metrics and presents them with meaningful names.
}
\details{
The function extracts data for the given `sample_id` and selects a subset of columns related to immune signatures. These columns are renamed for better readability in the plot. The data is then transformed into a long format suitable for plotting with `ggplot2`. The bar plot visualizes the immune pathway or signature scores, differentiated by the analysis type (Average or GSVA) in separate facets.
}
\examples{
# se is a pre-loaded SummarizedExperiment object
sample_id <- "sample123"
plot_immune_signature_one_sample(se, sample_id)

}
