% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ovrseq_results.R
\name{plot_quantiseq_one_sample}
\alias{plot_quantiseq_one_sample}
\title{Plot QuanTIseq Metrics for a Specific Sample}
\usage{
plot_quantiseq_one_sample(se, sample_id)
}
\arguments{
\item{se}{A `SummarizedExperiment` object containing quantiseq-related metrics in its `colData`.}

\item{sample_id}{A character string specifying the ID of the sample to be plotted.}
}
\value{
A `ggplot` object representing the bar plot of QuanTIseq metrics for the specified sample.
}
\description{
This function creates a bar plot of QuanTIseq metrics for a specified sample from a `SummarizedExperiment` object. It filters and plots only those metrics that are related to QuanTIseq.
}
\details{
The function first checks if the specified `sample_id` is present in the `colData` of the `SummarizedExperiment` object. It then filters the data to include only those columns that contain the term "quantiseq" (case-insensitive). A bar plot is created using this filtered data, providing a visual representation of the QuanTIseq metrics for the chosen sample.
}
\examples{
# se is a pre-loaded SummarizedExperiment object
sample_id <- "sample123"
plot_quantiseq_one_sample(se, sample_id)

}
