% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerabilitymap.R
\name{plot_vulnerabilitymap}
\alias{plot_vulnerabilitymap}
\title{Plot Vulnerability Map}
\usage{
plot_vulnerabilitymap(se)
}
\arguments{
\item{se}{A SummarizedExperiment object containing the necessary data for plotting. It must include
BRCAness probabilities and CYT to C1QA ratios, or data to compute them.}
}
\value{
A ggplot object representing the vulnerability map.
}
\description{
This function creates a vulnerability map based on three key variables: the vulnerability score,
the BRCAness score, and the cytolytic activity (CYT) to C1QA ratio (C2C). The BRCAness score is
derived from the prediction probability of a random forest classifier. The C2C ratio is calculated
using the log2-transformed expression values of GZMB, PRF1, and C1QA.
}
\details{
The vulnerability score is defined as a weighted sum of the BRCAness probability and the C2C ratio,
where the weights are determined by a logistic regression model. This model uses log2 intensity
expression values and SigMA status (mutational signature 3) data from the TOPACIO cohort with the
treatment response as a binary dependent variable.

The function plots a two-dimensional map with the transformed C2C ratio and BRCAness probability as
coordinates. The map is color-coded based on the vulnerability score.


C2C is computed as follows:
\deqn{C2C = 0.5 \times (\log_2(GZMB + 1) + \log_2(PRF1 + 1)) / \log_2(C1QA + 1)}

The vulnerability score is computed using:
\deqn{Vulnerability Score = 2.687 \times BRCAness Probability + 1.051 \times C2C}

For visualization, C2C values are transformed to a range between 0 and 1 using a sigmoid function.
The BRCAness probability and the transformed C2C values are used as coordinates on the map.
}
\examples{
# se is a SummarizedExperiment object with required data
plot_vulnerabilitymap(se)

}
