% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{RPKMNorm}
\alias{RPKMNorm}
\title{RPKM normalization of count data}
\usage{
RPKMNorm(count_data)
}
\arguments{
\item{count_data}{A matrix of count data with genes in rows and samples in columns.}
}
\value{
A RPKM-normalized matrix with the same dimensions as \code{count_data}.
}
\description{
Takes a count data matrix, and applies RPKM normalization.
}
\examples{
# Load example data from SummarizedExperiment package
data("example_se")
# Extract count data and gene length data
counts <- assay(example_se)
# Apply RPKM normalization
normalized_counts <- RPKMNorm(counts)

}
