% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tcgaStats}
\alias{tcgaStats}
\title{Internal tcgaStats Data}
\format{
A data frame with x rows and y columns.
}
\description{
Description of the tcgaStats data and its structure. This data is for internal package use.
}
\keyword{internal}
