% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{TPMNorm}
\alias{TPMNorm}
\title{TPM normalization of count data}
\usage{
TPMNorm(count_data, gene_length)
}
\arguments{
\item{count_data}{A matrix of count data with genes in rows and samples in columns.}

\item{gene_length}{A vector with gene IDs as names and transcript lengths in kilobases (KB).}
}
\value{
A TPM-normalized matrix with the same dimensions as \code{count_data}.
}
\description{
Takes a count data matrix and a gene length data frame, and applies TPM normalization.
}
\examples{
# Load example data from SummarizedExperiment package
data("example_se")
# Extract count data and gene length data
counts <- assay(example_se)
gene_length <- rowData(example_se)$gene_length
# Apply TPM normalization
normalized_counts <- TPMNorm(counts, gene_length)

}
