% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brcaness.R
\name{train_rf}
\alias{train_rf}
\title{Train a random forest model on gene expression data}
\usage{
train_rf(se, label, gene_signature)
}
\arguments{
\item{se}{A SummarizedExperiment object containing gene expression data}

\item{label}{A string indicating the name of the column in colData containing the label to use as the target variable for classification}

\item{gene_signature}{A character vector containing the names of genes to use in the analysis}
}
\value{
A trained random forest model
}
\description{
This function trains a random forest model on a summarized experiment object
containing gene expression data, using a specified gene signature to select a
subset of genes, and a specified label in the colData to use as the target variable.
The function returns the best random forest model.
}
\examples{
# Load the TCGA_OV.rda and BRCAness_signature.rda files
data(TCGA_OV)
data("brcaness_signature")

# Train a random forest model using the BRCAness label
rf_model <- train_rf(se, "BRCAness", brcaness_signature)

# Print the model
print(rf_model)

# Make predictions using the model
predictions <- predict(rf_model, newdata = t(assay(TCGA_OV)[rownames(BRCAness_signature),]))

# Print the predictions
print(predictions)
}
