% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration_status.R
\name{train_rf_classifier_infiltration_status}
\alias{train_rf_classifier_infiltration_status}
\title{Train Random Forest Classifier for Tumor Infiltration Status}
\usage{
train_rf_classifier_infiltration_status(
  se,
  gene_signature,
  num_estimators = 300
)
}
\arguments{
\item{se}{A SummarizedExperiment object containing gene expression data and a column `TinfStatus` in its `colData`
which indicates the tumor immune phenotype.}

\item{gene_signature}{A character vector containing the names of genes in the signature to be used for model training.}

\item{num_estimators}{The number of trees to grow in the random forest model. Default is 300.}
}
\value{
A random forest model object trained on the gene expression data and tumor immune phenotype status.
}
\description{
Trains a random forest classifier using a predefined gene signature to predict tumor immune phenotypes,
such as infiltrate, excluded, or desert based on the gene expression profiles from a SummarizedExperiment object.
}
\details{
The function extracts the expression data for the genes in the provided signature from a SummarizedExperiment object
and uses the `TinfStatus` from the column metadata to train a random forest classifier.
The `TinfStatus` should be a factor that represents the tumor immune phenotype status with levels such as
'infiltrate', 'excluded', or 'desert'. The gene expression matrix is transposed to ensure proper dimensions
for the random forest training function. The `train` function from the `caret` package is used for training
the model, with the number of trees set by the `num_estimators` parameter.

The function sets the seed for reproducibility and ensures the output is a trained model
that can predict the `TinfStatus` based on gene expression profiles.
}
\examples{
# Assuming 'se' is a preloaded SummarizedExperiment object with the required data
# and 'tumor_immune_phenotype_signature' is the predefined gene signature list:
rf_classifier <- train_rf_classifier_infiltration_status(se, tumor_immune_phenotype_signature)

}
\references{
Desbois M, Udyavar AR, Ryner L, Kozlowski C, Guan Y, Dürrbaum M, et al. Integrated digital
pathology and transcriptome analysis identifies molecular mediators of T-cell exclusion in
ovarian cancer. Nat Commun. 2020;11:5583.
}
