% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumor_immune_phenotype_signature}
\alias{tumor_immune_phenotype_signature}
\title{Tumor Immune Phenotype Signature}
\format{
A list containing gene symbols representing the 148 genes in the tumor immune phenotype signature.
}
\usage{
data(tumor_immune_phenotype_signature)
}
\description{
The `tumor_immune_phenotype_signature` object contains a gene signature that classifies
ovarian cancer tumor immune phenotypes. It was developed using integrated digital pathology
and transcriptome analysis, with a focus on CD8+ T cell presence and position within the tumor.
}
\details{
A gene signature developed based on digital pathology and transcriptome analysis
to classify tumor immune phenotypes (infiltrate, excluded, desert) in ovarian cancer.
This signature was derived from a classification model using 157 genes that describe
the presence and position of CD8+ T cells relative to the tumor center or margin
in the ICON7 cohort.
}
\examples{
# Load the tumor immune phenotype signature
data(tumor_immune_phenotype_signature)

# Access the gene symbols in the signature
signature_genes <- tumor_immune_phenotype_signature

}
\references{
Desbois M, Udyavar AR, Ryner L, Kozlowski C, Guan Y, Dürrbaum M, et al. Integrated digital
pathology and transcriptome analysis identifies molecular mediators of T-cell exclusion in
ovarian cancer. Nat Commun. 2020;11:5583.
}
