% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{update_se_with_entrez_ids}
\alias{update_se_with_entrez_ids}
\title{Update Gene Symbols to Entrez IDs in a SummarizedExperiment}
\usage{
update_se_with_entrez_ids(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A SummarizedExperiment object with Entrez IDs as row names.
}
\description{
This function takes a SummarizedExperiment object with gene symbols and updates the row names with corresponding Entrez IDs.
}
\examples{
# Assuming you have a SummarizedExperiment object named 'se'
se_updated <- update_se_with_entrez_ids(se)
rownames(se_updated)
}
