% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPS_genes}
\alias{IPS_genes}
\title{Immunophenoscore (IPS) Genes and Weights}
\format{
A data frame with the following columns:
\describe{
  \item{GENE}{(character) Official gene symbol.}
  \item{WEIGHT}{(numeric) Weight of the gene in the IPS calculation.}
  \item{CATEGORY}{(character) The immunological category to which the gene belongs
  (e.g., MHC, CP, EC, SC, MDSC, TREG, AZ).}
}
}
\source{
Immunophenogram project: https://github.com/icbi-lab/Immunophenogram
}
\usage{
data(IPS_genes)
}
\description{
A dataset containing genes and corresponding weights used to calculate the
Immunophenoscore (IPS), which is a measure of the immune landscape of a tumor.
}
\examples{
# Load the IPS genes and weights
data(IPS_genes)

# Access the gene symbols and weights
ips_genes <- IPS_genes$GENE
ips_weights <- IPS_genes$WEIGHT

}
\references{

}
