% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{load_brcaness_classifier}
\alias{load_brcaness_classifier}
\title{Load BRCAness Classifier}
\usage{
load_brcaness_classifier()
}
\value{
The pre-trained BRCAness classifier as a list.
}
\description{
Loads the pre-trained BRCAness classifier from the "brcaness_classifier.rda"
data file and returns it as the `brcaness_classifier` object.
}
\examples{
# Load the BRCAness classifier
brcaness_classifier <- load_brcaness_classifier()

}
