% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OvaRSeqDataSet.R
\name{OvaRSeqDataSet}
\alias{OvaRSeqDataSet}
\title{OvaRSeq constructor}
\usage{
OvaRSeqDataSet(countData, colData)
}
\arguments{
\item{countData}{a matrix or data frame contains gene count}

\item{colData}{a \code{DataFrame} or \code{data.frame}}

\item{...}{optional arguments passed to \code{SummarizedExperiment}}
}
\value{
a OvaRSeq object
}
\description{
OvaRSeq constructor
}
